+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 The format here is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3] - 2023-08-13
+### Added
+- `DayInWeek` now provides for expressions such as "every Monday",
+  "every Tuesday and Thursday", and even "every second Friday". #4
+
+### Fixed
+- Fixed broken handling of start and end in same month for
+  `RangeEveryYear`. This also fixes single day expressions, such as
+  "every 1 September". #5
+
+- Updated Surefire version so that JUnit 5 tests are actually run
+  during the build. #6
+
 ## [0.2] - 2023-07-09
 ### Changed
 - Various improvements to encapsulation, including: making fields